requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-rest-auth==0.9.5
django-rq==2.3.2
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
rq==1.4.2
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
